requirements.txt
beautifulsoup4
requests
streamlit
google-generativeai
python-dotenv
beautifulsoup4
requests
